created the checks class
