+++ src/github/pullRequest.ts
@@ -111,55 +111,4 @@ export class PullRequestApi {
   getAuthor(): string {
     return this.pr.user.login;
   }
-
-  /**
-   * Generates a Check Run or modifies the existing one.
-   * This way we can aggregate all the results from different causes into a single one
-   * {@link https://docs.github.com/en/rest/checks/runs?apiVersion=2022-11-28}
-   * @param checkResult a CheckData object with the final conclussion of action and the output text
-   * {@link CheckData}
-   */
-  async generateCheckRun(checkResult: CheckData): Promise<void> {
-    const checkData = {
-      ...checkResult,
-      owner: this.repoInfo.owner,
-      repo: this.repoInfo.repo,
-      external_id: "review-bot",
-      head_sha: this.pr.head.sha,
-      name: "review-bot",
-    };
-
-    const { data } = await this.api.rest.checks.listForRef({
-      owner: this.repoInfo.owner,
-      repo: this.repoInfo.repo,
-      ref: this.pr.head.sha,
-    });
-
-    this.logger.debug(`Searching for a match for id ${checkData.external_id}. Found ${data.total_count} checks`);
-
-    for (const check of data.check_runs) {
-      if (check.external_id === checkData.external_id) {
-        this.logger.debug(`Found match: ${JSON.stringify(check)}`);
-        await this.api.rest.checks.update({ ...checkData, check_run_id: check.id });
-        this.logger.debug("Updated check data");
-        return;
-      }
-    }
-
-    this.logger.debug("Did not find any matching status check. Creating a new one");
-
-    const check = await this.api.rest.checks.create(checkData);
-
-    // We publish it in the action summary
-    await summary
-      .emptyBuffer()
-      .addHeading(checkResult.output.title)
-      // We redirect to the check as it can changed if it is triggered again
-      .addLink("Find the result here", check.data.html_url ?? "")
-      .addBreak()
-      .addRaw(checkResult.output.text)
-      .write();
-
-    this.logger.debug(JSON.stringify(check.data));
-  }
 }
